pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "imagekitio>=5.2.0",
    "uvicorn[standard]>=0.41.0",
]
